+++ src/lib/utils/validations/validateCircleProps.ts
