+++ src/lib/utils/validations/validateScrollValue.ts
@@ -1,4 +1,4 @@
-import { Coors } from "../types";
+import { Coors } from "../../types";
 
 const validateScrollValue = (scrollValue: Coors | undefined) => {
   if (typeof scrollValue == "undefined")
